Added more to home page
Btw this was supposed to be a part of the previous commit, but I accidentally commited from a terminal that was open in the public/skill-icons, bacause I was doing some stuff there. The stuff was doanloading all of the icons from the terminal, which I actually didn't do, because it'd take a long time. Instead I downloaded em with a bash file.