Starting from scratch
I figured out, that the codebase of my previous portfolio (which I copied a lot from into this new one) was sooo unclean, that I'm rather going to start from scratch, with a clean NextJS project. But I'ma keep the github repo, because I'm to lazy to set everything up once more. Also, the only thing I kept are the css accent colors and the homepage component, which I deleted everything from anyways.